Modification to continue the test without starting from zero. Reading the previous file to know the previous state, writting after.
Reviewed-on: http://gerrit.aldebaran.lan/46927
Approved: enalepa <enalepa@aldebaran-robotics.com>
Tested-by: enalepa <enalepa@aldebaran-robotics.com>

+++ test_cables_routing/conftest.py
@@ -3,10 +3,12 @@ import subdevice
 import time
 import threading
 import tools
+import os
+import csv
 
 
 @pytest.fixture(scope="session")
-def cables_routing(request, dcm, mem):
+def cables_routing(request, dcm, mem, rest_pos):
     """
     Fixture which log number of cables routing
     """
@@ -19,9 +21,32 @@ def cables_routing(request, dcm, mem):
 
         parameters = tools.read_section("test.cfg", "CablesRoutingParameters")
 
-        passage_de_cables = 0
-        data = open(parameters["Log_file_name"][0], 'w')
-        data.write("CablesRouting,Time\n")
+        # If file already exists, reading to know the previous state
+        if os.path.exists(parameters["Log_file_name"][0]):
+            data = open(parameters["Log_file_name"][0], 'r')
+            csv_reader = csv.reader(data)
+            variables_list = []
+            row_num = 0
+            for row in csv_reader:
+                data_num = 0
+                if row_num == 0:
+                    for value in row:
+                        variables_list.append([value])
+                else:
+                    for value in row:
+                        variables_list[data_num].append(value)
+                        data_num += 1
+                row_num += 1
+            passage_de_cables = int(variables_list[0][row_num - 1])
+            precedent_time = float(variables_list[1][row_num - 1])
+            data.close()
+            data = open(parameters["Log_file_name"][0], 'a')
+
+        else:
+            data = open(parameters["Log_file_name"][0], 'w')
+            data.write("CablesRouting,Time\n")
+            passage_de_cables = 0
+            precedent_time = 0.
 
         time_init = time.time()
         while not thread_flag.is_set():
@@ -31,7 +56,7 @@ def cables_routing(request, dcm, mem):
                     gyro_y.value > float(parameters["Maximum"][0]):
                 passage_de_cables += 1
                 data.write(str(passage_de_cables) + "," +\
-                        str(time.time() - time_init) + "\n")
+                        str(time.time() - time_init + precedent_time) + "\n")
                 data.flush()
                 time.sleep(2)
